Added a new tab for the Project Spaces.
TODO: when returning the share object, put a new attribute on it to remark if the share is from a project space or not.
The solution to guess if a share is from a project space is checking if the path of the folder is inside the eos_project_prefix.