[Plugin|Patch] Notebook viewer app integration
IPython/Jupyter notebooks have "text/plain" as mimetype. OC uses files
mimetype to trigger the application. So, to be able to open the notebook
with the nb_viewer instead the texteditor, we have to fake the mimetype
in the client.
Alternative: Ask OC/Pull Request to register application by both,
mimetypes and file extension