[Patch] Sharing file by link UI update bug fix
When sharing a file by id, the client will request the share request to
refresh the info in the user interface. It will compare the
stored-in-client file id and the one received to update the properly
one. As we do not use stable file id, a check by eos path have been
added