[Patch] Fix for user share list in "Share with ..." tabs
ownCloud create a new object for the details view in every tab,
althought the share the same ID in html. This was causing jQuery to mess
the object returned when selecting them via ID. This patch change it to
use the object's internal html object to handle all UI updates.