OCREPORT: allow user backend to get emails
Without this patch to get the mail address of a user, the email has to
be registered in the settings of the user, i.e. in the database.
This means at least 1 row per user in the database to get the email.
As we are using LDAP that offers the mail attribute we do not need to
have the information repeated.